add: #23 그린룸 관련 API 추가

+++ module-api/src/main/java/com/greenroom/moduleapi/service/greenroom/question/GreenRoomQuestionServiceImpl.java
@@ -9,11 +9,12 @@ import com.greenroom.modulecommon.exception.ApiException;
 import com.greenroom.modulecommon.repository.greenroom.question.GreenRoomQuestionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.List;
+import java.time.temporal.ChronoUnit;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.greenroom.modulecommon.constant.EntityConstant.Question.QUESTION_LENGTH;
@@ -23,6 +24,8 @@ import static com.greenroom.modulecommon.exception.EnumApiException.NOT_FOUND;
 @RequiredArgsConstructor
 @Service
 public class GreenRoomQuestionServiceImpl implements GreenRoomQuestionService {
+
+    public static final long MAX_EXPIRED_AT_INTERVAL = 24 * 60 * 60L;
 
     private final GreenRoomQuestionRepository greenRoomQuestionRepository;
     private final UserService userService;
@@ -40,7 +43,15 @@ public class GreenRoomQuestionServiceImpl implements GreenRoomQuestionService {
         User user = userService.getUser(userId);
         Category category = categoryService.getCategory(categoryId);
 
-        //FIXME: expiredAt 24시간이내 값인지 검증하기
+        if (expiredAt.isBefore(LocalDateTime.now())) {
+            throw new IllegalArgumentException("마감 기한은 현재시간보다 앞선 시간일 수 없습니다.");
+        }
+
+        long interval = ChronoUnit.SECONDS.between(expiredAt, LocalDateTime.now());
+
+        if (Math.abs(interval) > MAX_EXPIRED_AT_INTERVAL) {
+            throw new IllegalArgumentException("마감 기한은 현재시간보다 24시간을 초과할 수 없습니다.");
+        }
 
         GreenRoomQuestion greenRoomQuestion = GreenRoomQuestion.builder()
                 .question(question)
@@ -53,10 +64,10 @@ public class GreenRoomQuestionServiceImpl implements GreenRoomQuestionService {
     }
 
     @Override
-    public List<GreenRoomQuestion> getGreenRoomQuestions(Long userId, Pageable pageable) {
+    public Slice<GreenRoomQuestion> getMyGreenRoomQuestions(Long userId, Pageable pageable) {
         checkArgument(userId != null, "userId 값은 필수입니다.");
 
-        return greenRoomQuestionRepository.findAll(userId, pageable);
+        return greenRoomQuestionRepository.findMyGreenRoomQuestions(userId, pageable);
     }
 
     @Override
@@ -68,11 +79,26 @@ public class GreenRoomQuestionServiceImpl implements GreenRoomQuestionService {
     }
 
     @Override
+    public boolean isOwner(Long id, Long userId) {
+        checkArgument(id != null, "id 값은 필수입니다.");
+        checkArgument(userId != null, "userId 값은 필수입니다.");
+
+        GreenRoomQuestion greenRoomQuestion = getGreenRoomQuestion(id);
+
+        return greenRoomQuestion.getUser().map(user -> user.getId().equals(userId)).orElse(false);
+    }
+
+    @Override
     @Transactional
     public void delete(Long id) {
         checkArgument(id != null, "id 값은 필수입니다.");
 
         GreenRoomQuestion greenRoomQuestion = getGreenRoomQuestion(id);
+
+        if (greenRoomQuestion.getQuestionAnswers().size() > 0) {
+            throw new IllegalStateException("질문에 답변을 단 참여자가 있는 경우 삭제할 수 없습니다.");
+        }
+
         greenRoomQuestion.delete();
     }
 }
